timers: move mpl_slow_timer out of protocol_core

+++ stack/source/mpl/mpl.h
@@ -43,7 +43,7 @@ buffer_t *mpl_control_handler(buffer_t *buf, struct protocol_interface_info_entr
 
 bool mpl_forwarder_process_message(buffer_t *buf, mpl_domain_t *domain, bool decrement_hop_limit);
 
-void mpl_slow_timer(uint16_t seconds);
+void mpl_slow_timer(int seconds);
 
 void mpl_clear_realm_scope_seeds(struct protocol_interface_info_entry *cur);
